Refactor DXF versioning and text codec management
Switch DXF version handling from int to EoDxf::Version enum for improved type safety and clarity across EoDxfHeader, EoDxfReader, and EoTcTextCodec. Replace raw pointer with std::unique_ptr for EoTcConverter in EoTcTextCodec, and implement copy/move constructors and assignment operators for safe object management. Refactor converter creation into a static method, update SetCodePage and SetVersion to use modern C++ idioms, and clean up related code. These changes enhance memory safety, maintainability, and reduce risk of versioning bugs.

+++ Core Libraries/EoDxfLib/EoDxfHeader.cpp
@@ -82,7 +82,7 @@ void EoDxfHeader::ParseCode(int code, EoDxfReader* reader) {
       auto newVariant = std::make_unique<EoDxfGroupCodeValuesVariant>();
       m_currentVariant = newVariant.get();
       m_name = reader->GetString();
-      if (m_version < static_cast<int>(EoDxf::Version::AC1015) && m_name == "$DIMUNIT") { m_name = "$DIMLUNIT"; }
+      if (m_version < EoDxf::Version::AC1015 && m_name == "$DIMUNIT") { m_name = "$DIMLUNIT"; }
       m_variants[m_name] = std::move(newVariant);
     } break;
     case 10:
